Fix test suite compilation
Even though there are no unit tests so far, there has been a
test suite stanza in the Cabal configuration. However, the
Cabal configuration used `Hspec` and the dummy test module
`HUnit`. The dummy test module was updated to `Hspec` since
this is the testing framework that we are also using in the
free-compiler.