Rename `unit-tests` target
The `haskell-src-transformations-unit-tests` was renamed to
`unit-tests` since the contributing guidelines have to specify
how to run tests. Since there is a `tests` target alias in
Cabal. We can use the long name again. The long name has the
advantage, that it is less likely for a package with that
name to exist on Hackage.